+++ src/features/TreegeConsumer/TreegeConsumer.tsx
@@ -1,5 +1,8 @@
+import { LocalizationProvider } from "@mui/x-date-pickers";
+import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { Box, CircularProgress, ThemeOptions, ThemeProvider, useTheme } from "@tracktor/design-system";
+import dayjs from "dayjs";
 import type { CSSProperties } from "react";
 import OptionsProvider from "@/context/Options/OptionsProvider";
 import Standard from "@/features/TreegeConsumer/Standard";
@@ -8,6 +11,9 @@ import useTreegeConsumer, { OnSubmitReturn } from "@/features/TreegeConsumer/use
 import Headers from "@/types/Headers";
 import type TreeNode from "@/types/TreeNode";
 import { JsonFormValue } from "@/utils/formDataToJSON/formDataToJSON";
+import "dayjs/locale/fr";
+
+dayjs.locale("fr");
 
 export interface TreegeConsumerProps {
   /**
@@ -118,49 +124,51 @@ const TreegeConsumer = ({
   const queryClient = new QueryClient();
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <ThemeProvider theme={theme || themeProvider.palette.mode}>
-        {loading ? (
-          <Box display="flex" alignItems="center" justifyContent="center" height="100%">
-            <CircularProgress color="primary" />
-          </Box>
-        ) : (
-          <OptionsProvider options={options}>
-            {variant === "stepper" ? (
-              <Stepper
-                activeFieldIndex={activeFieldIndex}
-                firstFieldIndex={firstFieldIndex}
-                isLastField={isLastField}
-                style={style}
-                fields={fields}
-                readOnly={readOnly}
-                headers={headers}
-                fieldValues={fieldValues}
-                isLoadingFormValidation={isLoadingFormValidation}
-                handleChangeFormValue={handleChangeFormValue}
-                handlePrev={handlePrev}
-                handleSubmit={handleSubmit}
-                formCanBeSubmit={formCanBeSubmit}
-              />
-            ) : (
-              <Standard
-                fields={fields}
-                handleChangeFormValue={handleChangeFormValue}
-                handleSubmit={handleSubmit}
-                isLastField={isLastField}
-                readOnly={readOnly}
-                headers={headers}
-                fieldValues={fieldValues}
-                isLoadingFormValidation={isLoadingFormValidation}
-                style={style}
-                formCanBeSubmit={formCanBeSubmit}
-                ignoreFields={ignoreFields}
-              />
-            )}
-          </OptionsProvider>
-        )}
-      </ThemeProvider>
-    </QueryClientProvider>
+    <LocalizationProvider dateAdapter={AdapterDayjs} adapterLocale="fr">
+      <QueryClientProvider client={queryClient}>
+        <ThemeProvider theme={theme || themeProvider.palette.mode}>
+          {loading ? (
+            <Box display="flex" alignItems="center" justifyContent="center" height="100%">
+              <CircularProgress color="primary" />
+            </Box>
+          ) : (
+            <OptionsProvider options={options}>
+              {variant === "stepper" ? (
+                <Stepper
+                  activeFieldIndex={activeFieldIndex}
+                  firstFieldIndex={firstFieldIndex}
+                  isLastField={isLastField}
+                  style={style}
+                  fields={fields}
+                  readOnly={readOnly}
+                  headers={headers}
+                  fieldValues={fieldValues}
+                  isLoadingFormValidation={isLoadingFormValidation}
+                  handleChangeFormValue={handleChangeFormValue}
+                  handlePrev={handlePrev}
+                  handleSubmit={handleSubmit}
+                  formCanBeSubmit={formCanBeSubmit}
+                />
+              ) : (
+                <Standard
+                  fields={fields}
+                  handleChangeFormValue={handleChangeFormValue}
+                  handleSubmit={handleSubmit}
+                  isLastField={isLastField}
+                  readOnly={readOnly}
+                  headers={headers}
+                  fieldValues={fieldValues}
+                  isLoadingFormValidation={isLoadingFormValidation}
+                  style={style}
+                  formCanBeSubmit={formCanBeSubmit}
+                  ignoreFields={ignoreFields}
+                />
+              )}
+            </OptionsProvider>
+          )}
+        </ThemeProvider>
+      </QueryClientProvider>
+    </LocalizationProvider>
   );
 };